pyproject.toml
[project]
name = "langgraph-techwithtim-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain-openai>=1.1.7",
    "langchain[anthropic]>=1.2.7",
    "langgraph[openai]>=1.0.7",
    "python-dotenv>=1.2.1",
]
